Move ExitCode to the companion of StreamApp
