+++ core/jvm/src/main/scala/fs2/StreamApp.scala
@@ -8,6 +8,7 @@ import cats.implicits._
 
 import fs2.async.Ref
 import fs2.async.mutable.Signal
+import fs2.StreamApp.ExitCode
 
 abstract class StreamApp[F[_]](implicit F: Effect[F]) {
 
@@ -68,4 +69,14 @@ abstract class StreamApp[F[_]](implicit F: Effect[F]) {
 
   def main(args: Array[String]): Unit =
     sys.exit(doMain(args.toList).unsafeRunSync.code.toInt)
+}
+
+object StreamApp {
+  final case class ExitCode(code: Byte)
+
+  object ExitCode {
+    def fromInt(code: Int): ExitCode = ExitCode(code.toByte)
+    val success: ExitCode = ExitCode(0)
+    val error: ExitCode = ExitCode(1)
+  }
 }
